requirements.txt
asgiref==3.2.3
Django==3.0.3
pillow==7.0.0
pytz==2019.3
django-rest-auth==0.9.5
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
sqlparse==0.3.0